chore(DataSpreadsheet): refactored and created some custom hooks/utils
