+++ packages/cloud-cognitive/src/components/DataSpreadsheet/DataSpreadsheet.js
@@ -6,13 +6,7 @@
  */
 
 // Import portions of React that are needed.
-import React, {
-  useMemo,
-  useRef,
-  useEffect,
-  useState,
-  useCallback,
-} from 'react';
+import React, { useMemo, useRef, useState, useCallback } from 'react';
 import { useBlockLayout, useTable } from 'react-table';
 
 // Other standard imports.
@@ -23,12 +17,16 @@ import { getDevtoolsProps } from '../../global/js/utils/devtools';
 import { pkg } from '../../settings';
 import { getScrollbarWidth } from '../../global/js/utils/getScrollbarWidth';
 import { DataSpreadsheetBody } from './DataSpreadsheetBody';
-import { getCellSize } from './getCellSize';
+import { getCellSize } from './utils/getCellSize';
 import { DataSpreadsheetHeader } from './DataSpreadsheetHeader';
 import { useActiveElement } from '../../global/js/hooks';
-import { createActiveCellFn } from './createActiveCellFn';
+import { createActiveCellFn } from './utils/createActiveCellFn';
 import { deepCloneObject } from '../../global/js/utils/deepCloneObject';
 import { usePreviousValue } from '../../global/js/hooks';
+import { useResetSpreadsheetFocus } from './hooks/useResetSpreadsheetFocus';
+import { useSpreadsheetOutsideClick } from './hooks/useSpreadsheetOutsideClick';
+import { useMoveActiveCell } from './hooks/useMoveActiveCell';
+import { handleMultipleKeys } from './utils/handleMultipleKeys';
 import uuidv4 from '../../global/js/utils/uuidv4';
 // cspell:words rowcount colcount
 
@@ -62,6 +60,8 @@ export let DataSpreadsheet = React.forwardRef(
     },
     ref
   ) => {
+    const localRef = useRef();
+    const spreadsheetRef = ref || localRef;
     const focusedElement = useActiveElement();
     const [containerHasFocus, setContainerHasFocus] = useState(false);
     const [activeCellCoordinates, setActiveCellCoordinates] = useState(null);
@@ -99,27 +99,6 @@ export let DataSpreadsheet = React.forwardRef(
       useBlockLayout
     );
 
-    // Reset everything when spreadsheet loses focus
-    useEffect(() => {
-      if (
-        !focusedElement.classList.contains(
-          `${blockClass}--interactive-cell-element`
-        )
-      ) {
-        setContainerHasFocus(false);
-        removeActiveCell();
-        activeKeys.current = [];
-      }
-      if (
-        focusedElement.classList.contains(blockClass) ||
-        focusedElement.classList.contains(
-          `${blockClass}--interactive-cell-element`
-        )
-      ) {
-        setContainerHasFocus(true);
-      }
-    }, [focusedElement, removeActiveCell]);
-
     // Removes the active cell element
     const removeActiveCell = useCallback(() => {
       const activeCellHighlight = spreadsheetRef.current.querySelector(
@@ -149,31 +128,6 @@ export let DataSpreadsheet = React.forwardRef(
       },
       [spreadsheetRef]
     );
-
-    // Click outside useEffect
-    useEffect(() => {
-      const handleOutsideClick = (event) => {
-        if (
-          !spreadsheetRef.current ||
-          spreadsheetRef.current.contains(event.target) ||
-          event.target.classList.contains(
-            `${blockClass}__active-cell--highlight`
-          )
-        ) {
-          return;
-        }
-        setActiveCellCoordinates(null);
-        setSelectionAreas([]);
-        removeActiveCell();
-        removeCellSelections();
-        setContainerHasFocus(false);
-        activeKeys.current = [];
-      };
-      document.addEventListener('click', handleOutsideClick);
-      return () => {
-        document.removeEventListener('click', handleOutsideClick);
-      };
-    }, [spreadsheetRef, removeActiveCell, removeCellSelections]);
 
     const createActiveCell = useCallback(
       ({ placementElement, coords, addToHeader = false }) => {
@@ -219,6 +173,30 @@ export let DataSpreadsheet = React.forwardRef(
       ]
     );
 
+    useResetSpreadsheetFocus({
+      activeKeys,
+      focusedElement,
+      removeActiveCell,
+      setContainerHasFocus,
+    });
+
+    useSpreadsheetOutsideClick({
+      spreadsheetRef,
+      setActiveCellCoordinates,
+      setSelectionAreas,
+      removeActiveCell,
+      removeCellSelections,
+      setContainerHasFocus,
+      activeKeys,
+    });
+
+    useMoveActiveCell({
+      spreadsheetRef,
+      activeCellCoordinates,
+      containerHasFocus,
+      createActiveCell,
+    });
+
     const handleInitialArrowPress = useCallback(() => {
       // If activeCellCoordinates is null then we need to set an initial value
       // which will place the activeCell on the select all cell/button
@@ -252,81 +230,6 @@ export let DataSpreadsheet = React.forwardRef(
       },
       []
     );
-
-    const handleMultipleKeys = useCallback(() => {
-      const activeKeyValues = activeKeys.current;
-      const selectionAreasClone = deepCloneObject(selectionAreas);
-      const indexOfCurrentArea = selectionAreasClone.findIndex(
-        (item) => item.matcher === currentMatcher
-      );
-      const pointToUpdate = selectionAreasClone[indexOfCurrentArea]?.point2
-        ? selectionAreasClone[indexOfCurrentArea].point2
-        : selectionAreasClone[indexOfCurrentArea].point1;
-      // Down + Shift
-      if (
-        activeKeyValues.includes('Shift') &&
-        activeKeyValues.includes('ArrowDown')
-      ) {
-        if (rows.length - 1 === pointToUpdate.row) {
-          return;
-        }
-        const newPoint = {
-          row: pointToUpdate.row + 1,
-          column: pointToUpdate.column,
-        };
-        selectionAreasClone[indexOfCurrentArea].point2 = newPoint;
-        selectionAreasClone[indexOfCurrentArea].areaCreated = false;
-        setSelectionAreas(selectionAreasClone);
-      }
-      // Right + Shift
-      if (
-        activeKeyValues.includes('Shift') &&
-        activeKeyValues.includes('ArrowRight')
-      ) {
-        if (columns.length - 1 === pointToUpdate.column) {
-          return;
-        }
-        const newPoint = {
-          row: pointToUpdate.row,
-          column: pointToUpdate.column + 1,
-        };
-        selectionAreasClone[indexOfCurrentArea].point2 = newPoint;
-        selectionAreasClone[indexOfCurrentArea].areaCreated = false;
-        setSelectionAreas(selectionAreasClone);
-      }
-      // Up + Shift
-      if (
-        activeKeyValues.includes('Shift') &&
-        activeKeyValues.includes('ArrowUp')
-      ) {
-        if (pointToUpdate.row === 0) {
-          return;
-        }
-        const newPoint = {
-          row: pointToUpdate.row - 1,
-          column: pointToUpdate.column,
-        };
-        selectionAreasClone[indexOfCurrentArea].point2 = newPoint;
-        selectionAreasClone[indexOfCurrentArea].areaCreated = false;
-        setSelectionAreas(selectionAreasClone);
-      }
-      // Left + Shift
-      if (
-        activeKeyValues.includes('Shift') &&
-        activeKeyValues.includes('ArrowLeft')
-      ) {
-        if (pointToUpdate.column === 0) {
-          return;
-        }
-        const newPoint = {
-          row: pointToUpdate.row,
-          column: pointToUpdate.column - 1,
-        };
-        selectionAreasClone[indexOfCurrentArea].point2 = newPoint;
-        selectionAreasClone[indexOfCurrentArea].areaCreated = false;
-        setSelectionAreas(selectionAreasClone);
-      }
-    }, [selectionAreas, currentMatcher, columns, rows]);
 
     const handleKeyPress = useCallback(
       (event) => {
@@ -367,7 +270,14 @@ export let DataSpreadsheet = React.forwardRef(
           activeKeys.current = [...activeClone, key];
         }
         if (activeKeys.current?.length > 1) {
-          handleMultipleKeys();
+          handleMultipleKeys({
+            activeKeys,
+            selectionAreas,
+            currentMatcher,
+            rows,
+            setSelectionAreas,
+            columns,
+          });
         }
         // Allow arrow key navigation if there are less than two activeKeys OR
         // if one of the activeCellCoordinates is in a header position
@@ -484,13 +394,13 @@ export let DataSpreadsheet = React.forwardRef(
       [
         updateActiveCellCoordinates,
         handleInitialArrowPress,
-        handleMultipleKeys,
         activeCellCoordinates,
-        selectionAreas?.length,
         removeCellSelections,
         removeActiveCell,
-        columns.length,
-        rows.length,
+        columns,
+        rows,
+        currentMatcher,
+        selectionAreas,
       ]
     );
 
@@ -530,32 +440,6 @@ export let DataSpreadsheet = React.forwardRef(
       [removeCellSelections]
     );
 
-    // Adds active cell highlight to correct cell onKeyDown
-    useEffect(() => {
-      const activeCellPlacementElement = spreadsheetRef?.current.querySelector(
-        `[data-row-index="${activeCellCoordinates?.row}"][data-column-index="${activeCellCoordinates?.column}"]`
-      );
-      const shouldPlaceActiveCellInHeader =
-        activeCellCoordinates?.row === 'header' && true;
-      const selectAllElement = spreadsheetRef?.current.querySelector(
-        `[data-row-index="header"][data-column-index="header"]`
-      );
-      if (containerHasFocus) {
-        createActiveCell({
-          placementElement: activeCellCoordinates
-            ? activeCellPlacementElement
-            : selectAllElement,
-          coords: activeCellCoordinates,
-          addToHeader: shouldPlaceActiveCellInHeader,
-        });
-      }
-    }, [
-      activeCellCoordinates,
-      spreadsheetRef,
-      createActiveCell,
-      containerHasFocus,
-    ]);
-
     const handleKeyUp = (event) => {
       const { key } = event;
       // Remove key from active keys array on key up
@@ -568,8 +452,6 @@ export let DataSpreadsheet = React.forwardRef(
       }
     };
 
-    const localRef = useRef();
-    const spreadsheetRef = ref || localRef;
     return (
       <div
         {...rest}
