+++ packages/cloud-cognitive/src/components/DataSpreadsheet/utils/generateData.js
@@ -1,3 +1,10 @@
+/**
+ * Copyright IBM Corp. 2022, 2022
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
 // cspell:disable
 const pets = [
   'dog',
