+++ packages/cloud-cognitive/src/components/DataSpreadsheet/utils/getCellSize.js
